(WIP,UNTESTED) QENG-1600
This is intended to be groundwork for a rewrite of with_puppet_running_on that
would allow some level of heuristic determination as to whether or not the
service should be restarted. Presumably, this would allow us to skip restarts
for example when there have been no relevant configuration file changes.

I haven't touched any of this in quite a while so I don't know if it's in a
working state yet.